added test case and comments.
